separate fabfiles now with linting. For #68
