+++ .travis.yml
@@ -11,4 +11,6 @@ install:
 script:
 - pep8 lib
 - pyflakes lib
+- pep8 fabfile
+- pyflakes fabfile
 - coverage run setup.py test
